% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution des trajets par jour de la semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets vélos}
}
\value{
Un data.frame avec le nombre de trajets par jour de la semaine
}
\description{
Calculer la distribution des trajets par jour de la semaine
}
