% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les anomalies du jeu de données}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets vélos}
}
\value{
Un data.frame sans les lignes anormales
}
\description{
Filtrer les anomalies du jeu de données
}
