% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par numéro de boucle}
\usage{
filtrer_trajet(trajet, boucle = NULL)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets vélos}

\item{boucle}{Un vecteur de numéros de boucle à sélectionner}
}
\value{
Un data.frame filtré selon les boucles choisies
}
\description{
Filtrer les trajets par numéro de boucle
}
