% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Visualiser la distribution des trajets par jour de la semaine}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets vélos}
}
\value{
Un graphique ggplot2 en colonnes
}
\description{
Visualiser la distribution des trajets par jour de la semaine
}
