% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver la paire boucle-jour avec le maximum de passages}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les trajets vélos}
}
\value{
Un data.frame avec la boucle, le jour, le total et les moyennes
}
\description{
Trouver la paire boucle-jour avec le maximum de passages
}
